Use CircleCI, and stop using Docker Compose
We don't need Compose as we can just use Docker with the volume option.

Travis was erroring out because it uses a volume mount, and Jekyll made
a change to run commands as user 'jekyll' rather than root.
See https://github.com/jekyll/docker/issues/116

With this Circle setup, we don't need a volume mount. There's probably a
way to do it with Travis too, but just switching is easier.