Handle updating a post
If a post has been updated, display the date it was updated, along with
the date it was originally published. Hugo automtically updates the
lastmod value in the sitemap, so we don't have to worry about that.

For crawlers, add HTML time tags, along with "datePublished" and
"dateModified" microdata attributes. Currently, Google isn't displaying
the dates of the posts in the search results. Maybe these changes will
fix that.